Raise a better error for pages without a H1 tag
It often happens that people forget to add a h1 tag. The headings
parsing code can't handle this and crashes. Helpfully, it will tell the
user:

```
undefined method `depth' for nil:NilClass
```

This commit changes the behaviour to raise a more informative error,
namely:

```
No H1 tag found. You have to at least add one H1 heading to the page.
```

Even better would be to avoid the error altogether, but I'm having a
lot of trouble understanding the recursive nature of
`HeadingTreeBuilder`.