Implement an optional "contribution banner"
This implements an optional new feature, the "contribution banner".
It's a small banner under the content of the page that links to:

- The source of the current page on GitHub, so that people can easily
propose changes to the content
- A pre-filled GitHub issue, so that people can notify the owners of
problems with the content
- The GitHub repository

The feature is optional and needs to be turned on explicitly in
`tech-docs.yml`.

This is extracted from the GOV.UK developer docs, where it has been in
use for around a year.

Part of the Q4 GOV.UK 🔥 firebreak: https://trello.com/c/QbspvRc2