Depend on GDS's fork of middleman search
We've been depending on this fork for a while, but projects have had
to include a line in their Gemfile to point to the github repo,
because it's not possible for a gem to depend on another gem that
hasn't been packaged.

Now that we've published our fork to rubygems, we can depend on it
directly.