Fix unclickable menu etc when there is no TOC
We don't need the 'sticky TOC offset' when there is no TOC.
This removes that when there is no sidebar and therefore no TOC.

This fixes that the Menu isn't clickable on mobile
when the TOC is not there.
It also fixes links above headings not being clickable,
[https://github.com/alphagov/tech-docs-template/issues/123],
but only in the case when there is no sidebar.