Automate releases to Rubygems
This sets up a deployment on Travis, which hopefully means it'll
release a new version every time we merge a new version to master. This
prevents the project from being dependent on me to release new versions.

We've encrypted the Rubygems API key. The API key belongs to the
https://rubygems.org/profiles/govuk user. Note that this user belongs
to the GOV.UK programme, which has no formal relationship to the tech
docs template. However until there's a team officially owning the
project, we'll have to do it like this.

Thanks to @36degrees for helping me understand Travis.