Fix unclickable links
Bug: When links are in the last line above a heading,
they are currently not clickable.

Feature: Following a link to an anchor on mobile currently
appears below the sticky table of contents.

But that feature also causes the unclickable links bug,
which is why this removes that feature again.

Therefore this intentionally re-introduces the lesser evil
that headings will be hidden underneath the TOC on mobile
to fix alphagov/tech-docs-template#123.