Firebreak: Add show_expiry global option
Sometimes we want to know when the page was last reviewed and when it's
due for review again (via the Manual Spaniel, for example) but we don't
want users to see that the page has expired. The `show_expiry` config
option allows you to hide the review date and expired banner from all
pages.