Allow search results to be opened in new tabs
Fixes an issue whereby if you 'CMD + Click' or 'Right-Click > Open in New Tab'
on search results, the search results are closed. (This made it difficult to
line up several useful looking results in new tabs; you'd either have to click
on the search input again and re-trigger the search to re-show the results and
enable opening a result in a new tab, or give up on tabs altogether and navigate
in your current window.
    
The selector '.toc' contains the search input, the sidebar and the search results
themselves. So calling 'hideResults' when '.toc a' is clicked meant that the
search results were hidden the moment you click on a search result.
    
The selector '#toc' contains only the search input and sidebar - NOT the search
results, so we can safely call 'hideResults' when '#toc a' is clicked. This means
we can right-click on search results and open them in new tabs, without the results
being cleared. It also means we haven't broken the existing behaviour (which is that
if a navigation link in the sidebar is clicked, the search results should be hidden - as
the search results are overlayed on top).