Add GOV.UK Frontend CSS to this project
This adds the CSS from GOV.UK Frontend to this project. We do this by
adding the library to package.json. Whenever we publish the gem on
Travis, we first do an `npm install`, which will install the library in
the `node_modules` directory, which is then bundled into the gem (as
determined by `spec.files` in the gemspec.