Ensure GOV.UK frontend is installed via gemspec
The .travis.yml does this but putting it in the gemspec means it happens
locally as well when running 'rake build'

If this gem, when built does not include govuk-frontend, then it
functionally does not work. Add a check in the gemspec that a file from
govuk-frontend exists, so that building the gem will fail if 'npm
install' has not been run, or if govuk-frontend has been removed.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>
Co-authored-by: Jon Heslop <jon.heslop@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>