Increase line-height for <code> in <td>s
Text in our `<code>` elements is larger than the
default Transport font on OSX, due to it using
the Monaco font. This creates a taller formatting
box for each line of text it has.

We set a smaller font-size, and line-height in
`<td>`s. This works for Transport but text
in `<code>` sections appears squashed vertically.

This sets the line-height for `<code>` to match
that of content outside tables, to stop it
squashing up.

This has been tested with Consolas (the font
Windows uses for `<code>`) and all other
fallbacks.