Improve table of contents accessibility
The table of contents fails WCAG 2.4.3. When the table of contents is
expanded, a search field and a list of (sometimes) collapsible options
appears. Focus is automatically set to the first link on the list, which
means that screen reader users might miss both the search box and the
"hide table of contents" button.
The changes introduced in this PR ensure that focus is not set on the
first nav option, but on the entire region. This way when the user
navigates forward, they are able to encounter the search box and the
"close" button.

Additionally, adjust the markup for the open/close TOC buttons to use
the <button> element instead of an anchor tag. The reason is that these
elements both look and behave as buttons, not as links. This should aid
assistive tech users in navigating this functionality (for ex using
voice control)