[WIP] Change to using Kramdown
A quick spike to see if we could easily switch the markdown processor
from redcarpet to kramdown and get some tech docs rendered.

This was sucessful when checked using the govuk-account-team-manual
source and generally was us following along from a [similar pr][pr] in
the govuk-tech-docs repo.

Not sure if we want to persue this, or if we want to focus on switching
the renderer to be GOVspeak.

[pr]: https://github.com/alphagov/govuk-developer-docs/pull/2785/files#diff-80046815adcd86408da51ac00b9a94cf32ec30f97d58da320fe1a0711c3f0da4