Switch from Redcarpet to Kramdown
We recently received a [report][0] from Digital Accessibility Centre (a
specialist accessibility testing organisation based in Wales). This
report uncovered some [accessibility issues][1] that we need to fix.

One of these issues is that some tables in the content do not have
column row headers when needed - the first column on tables with 3 or
more columns does not have the `<th>` element. This fails
[WCAG 2.1 success criterion 1.3.1 (Info and Relationships)][2]. Changing
the markdown parser from Redcarpet to Kramdown fixes this issue.

In a [previous tech-docs-template commit][3], we had made the opposite
change and moved from Kramdown to Redcarpet because Kramdown didn't
support fenced code blocks. This isn't an issue now because we are using
the GitHub-flavoured markdown (enforced by the line `input: "GFM"`),
which does support fenced code blocks.

[0]: https://trello-attachments.s3.amazonaws.com/5ebbaf9d9e228c4bf72cd106/6037b26548297f5f44d61173/5a6ca3f5267d8d303c2b3c2ac2cc5aa2/DAC_Accessibility_Report_for_Cabinet_Office_GOV.UK_Accounts_WCAG_2.1_Feb_2021.pdf
[1]: https://docs.google.com/spreadsheets/d/1s53eGMBj2hdTVLRitwi673R-_uERWh0kOxOp9QsOVg0/edit#gid=0
[2]: https://www.w3.org/WAI/WCAG21/Understanding/info-and-relationships.html
[3]: https://github.com/alphagov/tech-docs-template/commit/3c0947c8cbc3bdfdda2181cc6ea63535f23978a9

Trello card: https://trello.com/c/vv0OR35g/630-march-update-tech-docs-template

Co-authored-by: Michael S. Walker <michael.s.walker@digital.cabinet-office.gov.uk>
Co-authored-by: Alan Gabbianelli <alan.gabbianelli@digital.cabinet-office.gov.uk>