Some search and keyboard navigation updates
Add a button to the search form as suggested in the DAC Acessibility
audit. Make input label more descriptive to help screen reader users
determine the purpose of the form.

Prevent tabbing beyond the TOC modal and/or search results when they are
open. Screen reader users should also not be able to access page content
that is not visible on the screen.