Use nokogiri to process html, instead of regex
This allows us to handle the more complex case where table headings
contain other markup (e.g. the heading might have an `<em>` tag, as per
the test).

It feels less hacky than using a regex, but it's quite a bit more code
so I'm not 100% sure whether it's an improvement.