Use new GOV.UK Frontend link styles
GOV.UK Frontend v3.12.0 comes with new (optional) link styles that are
designed to be easier to read and to have a clearer hover state [[1]].

This commit opts-in to the new link styles for the tech-docs-gem
stylesheet, and refactors the table of contents navigation and search
results links to use new the new link style mixins [[2]] so that they
are styled consistently with other links.

Note that this commit makes a small design change; search result links
are now underlined when not hovered.

[[1]]: https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#links-are-easier-to-read-and-have-a-clearer-hover-state-opt-in
[[2]]: https://frontend.design-system.service.gov.uk/sass-api-reference/#helpers-links