Update Node.js to latest 10.x.x release
Node.js 10.16.0 was released in April 2019, everyone should be using a
newer version than that in production, so there is no need to be using
it for development or testing.

This commit changes the .nvmrc so it no longer specifies a specific
minor or patch version of Node.js, just a specific major version. This
has the advantages for developers who can keep their local versions of
Node up to date without having to edit this file, and for Travis CI
which can use a cached version for quicker testing.