Refactor govuk-frontend Sass imports
govuk-frontend 3.7.0 changes the way that Sass dependencies work, so
that CSS compiles faster [[1]].

This commit changes the Sass file for the tech docs gem to take
advantage of the new system, by importing the component `index`
stylesheets which don't include dependencies multiple times.

This commit also changes the imports of govuk-frontend to import
`govuk/base` instead of importing the settings, tools, and helpers by
hand, as this will make it easier to migrate this repo to GOV.UK
Frontend v4.0.0 when it is released.

[1]: https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#sass-now-compiles-faster