Change search tests to use example app search index
This commit updates the Jasmine tests of the search functionality to
use the search index from the example site in this page, instead of a
"dummy" index.

This has the advantage of increasing the coverage of the tests, we can
now test that the `middleman-search` pipeline is configured the way we
want.

Note that the way we do this is by overriding the search module method
`downloadSearchIndex`. We could have added a variable on the module to
change the path to the search index, but then the `$.ajax` call would be
asynchronous, and then the tests won't pass over half of the time. It
would be nice to fix the tests to work with the default
`downloadSearchIndex` method, but I think that would require a large
refactor of both the tests and the implementation.