Fix format of search results
Adds a newline character after heading tags, so that the content in the
search index has a break between a heading and the following paragraph.

This fixes the search tests, but also changes the results users will
see, [the section on side effects below has more
details](#side-effects). This commit also changes the name of the search
tests slightly to be more accurate, see [the section on tests
below](#note-on-tests).

Description of issue
--------------------

Searching for a term that is in a header, the header and the following
`p` will appear in the search results with no space between them, for
instance, searching for "releases" in the GDS Way returns:

> … Low-risk releasesFrequent small changes make it easier to diagnose
problems, and the cost of fixing them is much lower than with large
releases …

This occurs because the page source has no whitespace between the header
tag and the paragraph tag, for example:

    <h3 id="low-risk-releases">Low-risk releases</h3><p>Frequent small ...

When `middleman-search` adds pages to the search index it uses
`Middleman::Sitemap::Resource#render` to render the template without
HTML tags [[1]]. This means the documents in the LunrJS data object will
include the content with no whitespace after heading text, for instance:

> Low-risk releasesFrequent small ...

Fix
---

To fix the issue, this commit changes the custom HTML renderer
`GOVUKTechDocs::TechDocsHtmlRender` in this repo to emit a newline
character after headings.

This does not change the formatting of pages in the browser, but the
change in the HTML source code propagates through the process above to
change the content in the LunrJS documents.

Side-effects
------------

This fix changes the behaviour of search results for the user.

A side effect of this change is that now a heading will be treated as a
single sentence by the search results splitter [[2]], and the content of
the following paragraph will no longer be included in the search
results. This may mean that search results are now less useful. For
instance, in the case where the page heading contains the search term, the
result will appear simply as the page title and the page heading (which
are usually the same). For instance, in the example for this repo,
searching for "expired" in the browser now shows the result:

> ### This is an expired page
>
> … This is an **expired** page …

If desired the function to render search results could be tweaked to
give more context after a sentence with a mark, but note that this would
affect all results, not just results in headings, as headings are not
distinguished from other sentences in the LunrJS documents.

Note on tests
-------------

The names of the search tests were not reflective of what the test was
testing, this commit also changes those names.

While investigating this bug I found that the search tests could be
improved in a number of ways; in fact I discovered it when updating the
tests, because the tests were out of sync with the actual behaviour of
the app. The tests are limited in their coverage, do not work
asynchronously (which is necessary for the search functionality), mix
setup and testing, and mock out an important part of the script.

However, to keep the scope of this PR limited, in this commit I just
address the fact that the test 'Search renders search results in the
search results area' is incorrectly named; in fact it is testing the
output of LunrJS.

To make the tests clearer, this commit renames the offending test and
reorganises the search tests to be more nested, so it is clearer which
method each test is testing.

[1]: https://github.com/alphagov/middleman-search/blob/f7addd7448cf49d3176022190b0aa1827c89fd83/lib/middleman-search/search-index-resource.rb#L134
[2]: https://github.com/alphagov/tech-docs-gem/blob/4f14a5fc541ae502432c36b5f24858571064f9d6/lib/assets/javascripts/_modules/search.js#L170