If page provided and heading level is 1, do not include the fragment
When clicking a nav link that is the main entry point for the page, the
focus should start at the very beginning. There is the "Skip to main content"
link to allow the user to bypass the nav.

Some nav examples do not include a page, for in page navigation. In this
case the fragment remains at level 1.