Remove IE8 fallback PNG
In 0e9ac92d we updated the header in line with changes made in GOV.UK Frontend v3.13.0, which involved changing the `<image>` tag to an `<img>` tag.

The path is incorrect (and was before), as the fallback PNG is actually located at `/assets/govuk/images/govuk-logotype-crown.png`.

Because of the change of tag (presumably) this is now causing downstream builds that have link checkers to fail, like in https://github.com/alphagov/govuk-frontend-docs/pull/139 (build log: https://app.travis-ci.com/github/alphagov/govuk-frontend-docs/builds/233880666).

We could fix the path, but given the layout is already completely broken in IE8 anyway, it seems like it makes more sense just to remove the fallback logic at this point.