Allow opting individual pages out of multipage nav
If `multipage_nav` is set in the frontmatter for the current page, use its value over the 'global' configuration when deciding whether to render a multi or single page table of contents.

This is most immediately useful for allowing us to add an example of 'single page navigation' to the existing example docs for testing purposes.