Add optional search page
Adds a search page that will be used for showing search results.

It is hidden from the navigation and from indexers (including itself),
to access it the browser uses the search input. If a browser access it
without a search query (either by visiting `/search` or pressing the
search button without a value in the search input) then a small help
message is shown, including a note about the requirement for JavaScript.

If the tech docs user does not include search in their site config then
this page will not be included in the site files.