Change search to only show results on search page
We want to remove the search results dialog and have a more traditional
form experience, as this is better for assistive tech.

This commit changes the search JavaScript so that:

- the search input is now progressively enhanced to submit the
query to the search page when the JavaScript loads.

- typing into the search input no longer causes results to appear
immediately.

- the search results are now rendered only on the search page `/search`.

- the search query can be loaded into the search input from the URL
query string.

- the text shown to the user when the search index is loading (which is
now more visible) is simpler to understand.

This commit also removes the search close button, as it is redundant.

Note that with this commit users can now get search analytics from the
URL and normal click events, so the custom Google Analytics code is less
needed; however this commit keeps it in place for backwards
compatibility.