Change search results to be normal flow page content
We want the search results on the search page to behave like normal page
content, instead of a dialog. Particularly, the skip link should take
the user to the search results, and other page furniture like the footer
should be visible and tab-accessible.

This commit moves the search results HTML from below the search input
into the search page template, in the normal content slot. The styles
are changed so that instead of being absolutely positioned and obscuring
the rest of the page the search results now push other contents down.
Note that the search help is not visible when search results are
visible, as the help is now hidden by the JavaScript function
`showResults()`.

This commit also has to copy some Sass from the `toc` stylesheet and
generally shuffle things around to get things looking the same as
before.