Change search input to be always visible on search page for mobile
Normally the search input is hidden behind the table of contents button
when viewing tech docs on a mobile device. However, for the search page,
we want the search input to be immediately visible, both so the user can
see their search query, and also so the help text makes sense.

This commit fiddles around with the styles so that on the search page,
the `html` element always has the class `has-search-results-open`, and
the table of contents pane is displayed so that the search box can be
displayed.

Note that this uses the `:not` psuedo-selector, which is not supported
in IE8; I may come back and refactor this later to try and
remove that requirement. In browsers that do not support `:not()` the
search input is always visible on mobile, which isn't a great hardship
however.