Fix mark styles in Windows High Contrast Mode
In some browsers (IE11, Chromium-based browsers) the marked text in
search results are unreadable with forced colour modes when the
background is black because the text colour for `<mark>` elements is
set to black.

This commit fixes this by re-asserting that the marked text should
be the same colour as the surrounding text.

We also have to set `background-color: Canvas`, because in Firefox
`background-color: transparent` is also ignored (unlike in IE11,
Chromiums).