Remove role alert from search results heading
Having role="alert" was stopping screenreaders including JAWS from
treating the search results heading as a heading.

We could have added the role to a span inside the heading, however after
checking guidance from MDN [1] I decided the alert role was not
appropriate for this situation, as aria-live="assertive" is sufficient
for screenreaders to be aware of changes to the text.

[1]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_alert_role