Change search results heading to h1
A page should always have a h1 heading. The search page has a h1
'Search' when the search results are not visible, but when the search
results are visible that is hidden, and the 'Results' heading is a h2.

This commit changes the heading 'Results' to be a h1.

This does mean that the page will have two h1s if the page loads without
CSS, so we also change the search results container to be hidden using
the 'hidden' HTML attribute, instead of CSS.