Change JavaScript to only download search index on search page
If the search JSON has poor cache headers (which sadly is probably quite
likely) then it'll be re-requested when the user hits the search page
anyway. If in future we make it easier to have better caching we can
always change it back.

To keep the tests working we add module functions `isOnSearchPage` and
`getQuery` which can be overridden.