Replace Travis CI with GitHub Actions
We want to use GitHub Actions as the tech docs template's continuous
integration/continuous deployment service. This commit removes
`.travis.yml` and adds equivalent workflows, `test.yaml` and
`publish.yaml`.

The script to publish to RubyGems was copied from the
discourse/publish-rubygems-action [[1]], with a modification to use the
github-actions[bot] user for tagging the release following suggestion from
fregante/setup-git-user action [[2]]. We avoid using the action directly
in accordance with the guidelines in the GDS Way [[3]].

[1]: https://github.com/discourse/publish-rubygems-action/blob/b55d7b9/entrypoint.sh
[2]: https://github.com/fregante/setup-git-user/blob/5ede2be/index.js
[3]: https://gds-way.cloudapps.digital/standards/source-code.html#using-github-actions-and-workflows

Use GitHub Actions bot account for creating release tags