Add accessible name to content pane div
When tabbing into the content pane with a screenreader we want
screenreaders that announce this to give a useful message to the user.

On VoiceOver and Chrome without this commit the announcement is
'and two items, group', after this commit the announcement is
'Content, group'. With NVDA and Firefox tabbing into the div is not
announced.