Pin Haml to versions before Haml 6
middleman-syntax does not support Haml 6 [[1]]. However, it does not
place any restriction on the version of Haml used, so bundler will
try and install the latest version of Haml. This commit adds a
restriction on the version of Haml in our gemspec. This is a bit of a
hack, as we don't actually require Haml ourselves, but it will make our
tests pass and provide a better experience for our users until
middleman-syntax is fixed.

[1]: https://github.com/middleman/middleman-syntax/issues/80