Add ruby-version to publish workflow
The ruby/setup-ruby action needs a version of Ruby specified to work.
Previously it was reading the `.ruby-version` file, however we removed
that in commit 2aa58cc. We could restore that file, but because
`ruby-install` doesn't support loose version specifiers, we'd have to
keep it maintained. Instead we just tell the action to install the
latest version of Ruby 3.