Bump govuk-frontend to v4.8.0
This release includes the ability to update the crown logo. You must do
this between 19 February and 1 March 2024. Note that many docs
repositories use this gem to set the version of the design system used.

We will need to carefully consider when to merge this and roll an update
as it comes with immediate operational impacts around use of favicons.

As we cannot use the Nunjuks template here as a feature flag, when we
are ready to enable the tudor crown we will need to follow the upgrading
guidance here:

https://github.com/alphagov/govuk-frontend/releases/tag/v4.8.0

Note that a v5.1 of the design system has also been published. I
suggest we get this merged then consider a broader set of upgrades as
part of any next major release to this gem