Add JS to prevent focused elements being obscured
The table of contents header, which you see on
smaller screens, is sticky and so can overlap
elements when they are focused because the browser
will only scroll to reveal elements when they are
off-screen. As far as the browser knows, the
header and the focused element are both onscreen.

This takes an approach from the paciello group
where we listen for focus events, check if the
header overlaps the element in focus and scrolls
to reveal it if so.