Move JS tests to use jasmine-browser-runner
The Jasmine ruby gem is deprecated:

https://github.com/jasmine/jasmine-gem?tab=readme-ov-file#discontinued

PhantomJS, the headless webkit browser used by the
Jasmine gem, is no longer being worked on:

https://phantomjs.org/

This follows the guidance in the Jasmine docs and
replaces the gem and phantomJS with the
jasmine-browser-runner npm package.

Because the new setup isn't in ruby, dependencies
'required' in JS files now need to be included
manually in the config.