Lock concurrent-ruby to 1.3.4
1.3.5 introduced a change that broke activesupport
(and so middleman, which uses it). It might be to
do with it dropping Logger as a dependency but
this is unclear. Full release notes are:

https://github.com/ruby-concurrency/concurrent-ruby/releases/tag/v1.3.5

The activesupport gem fixed this in v7-0-stable of
Rails:

https://github.com/rails/rails/pull/54264

So we're waiting for middleman to pull that
version in. They have an issue for this:

https://github.com/middleman/middleman/issues/2788