Update path to GOV.UK Frontend assets when using GOV.UK logo
These were pointing to the locations before the brand refresh,
so the wrong favicon were showing. Includes the favicons in the header,
as well as the Open Graph image.

Paths to the assets when not using the GOV.UK logo were left untouched,
as they are built from separate options.