Try to test copy of GOV.UK Frontend assets
In light of the recent issue with an asset not copied in the build,
I'm trying to add a test that GOV.UK Frontend's assets are copied.

For that, I'm looking to compare that all asset files in GOV.UK Frontend
are properly copied into the built site in an integration test.

However this doesn't seem to account for changes to the [`govuk_tech_docs.gemspec`][1],
which controls which files make up the assets of GOV.UK Frontend. For example, removing
one of the two globs in the `Dir` should make the test fail, but it doesn't.
Looking in the filesystem, all assets are properly copied regardless of the content of that string.

[1]: https://github.com/alphagov/tech-docs-gem/blob/e56addfeaa1b41a39505e23ff6ef9c477e2b6b84/govuk_tech_docs.gemspec#L28